Switching the chunks app over to being a South-managed app, for real
The chunks app was previously in a weird halfway between being a Django
ORM managed app and manually-created set of database schemas. This was
originally to be able to define foreign key constraints at the
schema-level and to create the tables as InnoDB instead of the
Django-default MyISAM.

Eventually, the chunk DB should not be directly modified by the chunk
loader, which renders the original design decision mostly null. That and
we should move to PostgreSQL as soon as possible.