Minor fixes / stop gaps
There is an error in notifications/models.py due to the multiple
authors change. I added a return statement right before the
offending line so that the bug doesn't effect the comment-posting UI.

Other fix is to fetch the reviewers for a submission when doing access control.
The Chunk.sorted_reviewers method does not include reviewers who are assigned
to the submission containing that chunk. This should probably be changed.